Use similar logic as before to display DOM
The logical flow of first adding items to projects, and then displaying
the list of projects, has remained the same. Except now, todo items
divs are appended as children to project divs which are appended as
children to the projectsContainer div. Next the former console based
display modules should be removed, and then users should be able to
add new projects and todo items in the webpage.