Add localStorage and update deleteTodoItem
The computer will check for listOfProjects in localStorage, if it finds
it it will load listOfProjects, if not listOfProjects = [];
Also, users are still able to delete individual todoItems. The
dependency on knowing a todoItem's project parent was removed to allow
localStorage to compile correctly (if was showing an error about
circular logic). The dependency was only there for the deleteTodoItem
module. That module was updated to rely less on dependencies and
accurately find the todoItem using logic